+++ src/components/theme/View/View.jsx
@@ -7,7 +7,12 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
-import { SummaryView, TabularView, DocumentView } from '../../../components';
+import {
+  SummaryView,
+  TabularView,
+  DocumentView,
+  ListingView,
+} from '../../../components';
 import { getContent } from '../../../actions';
 
 @connect(
@@ -99,6 +104,8 @@ export default class View extends Component {
         return <SummaryView content={this.props.content} />;
       case 'tabular_view':
         return <TabularView content={this.props.content} />;
+      case 'listing_view':
+        return <ListingView content={this.props.content} />;
       default:
         return <DocumentView content={this.props.content} />;
     }